+++ types/context.go
@@ -30,7 +30,9 @@ type Context struct {
 }
 
 // create a new context
-func NewContext(ms MultiStore, header abci.Header, isCheckTx bool, txBytes []byte, logger log.Logger) Context {
+func NewContext(ms MultiStore, header abci.Header, isCheckTx bool,
+	txBytes []byte, logger log.Logger, absentValidators []int32) Context {
+
 	c := Context{
 		Context: context.Background(),
 		pst:     newThePast(),
@@ -43,6 +45,7 @@ func NewContext(ms MultiStore, header abci.Header, isCheckTx bool, txBytes []byt
 	c = c.WithIsCheckTx(isCheckTx)
 	c = c.WithTxBytes(txBytes)
 	c = c.WithLogger(logger)
+	c = c.WithAbsentValidators(absentValidators)
 	return c
 }
 
@@ -127,6 +130,7 @@ const (
 	contextKeyIsCheckTx
 	contextKeyTxBytes
 	contextKeyLogger
+	contextKeyAbsentValidators
 )
 
 // NOTE: Do not expose MultiStore.
@@ -155,6 +159,9 @@ func (c Context) TxBytes() []byte {
 func (c Context) Logger() log.Logger {
 	return c.Value(contextKeyLogger).(log.Logger)
 }
+func (c Context) AbsentValidators() []int32 {
+	return c.Value(contextKeyAbsentValidators).([]int32)
+}
 func (c Context) WithMultiStore(ms MultiStore) Context {
 	return c.withValue(contextKeyMultiStore, ms)
 }
@@ -176,6 +183,9 @@ func (c Context) WithTxBytes(txBytes []byte) Context {
 }
 func (c Context) WithLogger(logger log.Logger) Context {
 	return c.withValue(contextKeyLogger, logger)
+}
+func (c Context) WithAbsentValidators(AbsentValidators []int32) Context {
+	return c.withValue(contextKeyAbsentValidators, AbsentValidators)
 }
 
 // Cache the multistore and return a new cached context. The cached context is