+++ x/stake/keeper.go
@@ -2,11 +2,13 @@ package stake
 
 import (
 	"bytes"
+	"sort"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/wire"
 	"github.com/cosmos/cosmos-sdk/x/bank"
 	abci "github.com/tendermint/abci/types"
+	crypto "github.com/tendermint/go-crypto"
 )
 
 // keeper of the staking store
@@ -34,28 +36,22 @@ func NewKeeper(cdc *wire.Codec, key sdk.StoreKey, ck bank.Keeper, codespace sdk.
 }
 
 // get the current in-block validator operation counter
-func (k Keeper) getCounter(ctx sdk.Context) int16 {
+func (k Keeper) getIntraTxCounter(ctx sdk.Context) int16 {
 	store := ctx.KVStore(k.storeKey)
-	b := store.Get(CounterKey)
+	b := store.Get(IntraTxCounterKey)
 	if b == nil {
 		return 0
 	}
 	var counter int16
-	err := k.cdc.UnmarshalBinary(b, &counter)
-	if err != nil {
-		panic(err)
-	}
+	k.cdc.MustUnmarshalBinary(b, &counter)
 	return counter
 }
 
 // set the current in-block validator operation counter
-func (k Keeper) setCounter(ctx sdk.Context, counter int16) {
+func (k Keeper) setIntraTxCounter(ctx sdk.Context, counter int16) {
 	store := ctx.KVStore(k.storeKey)
-	bz, err := k.cdc.MarshalBinary(counter)
-	if err != nil {
-		panic(err)
-	}
-	store.Set(CounterKey, bz)
+	bz := k.cdc.MustMarshalBinary(counter)
+	store.Set(IntraTxCounterKey, bz)
 }
 
 //_________________________________________________________________________
@@ -67,17 +63,14 @@ func (k Keeper) GetCandidate(ctx sdk.Context, addr sdk.Address) (candidate Candi
 	if b == nil {
 		return candidate, false
 	}
-	err := k.cdc.UnmarshalBinary(b, &candidate)
-	if err != nil {
-		panic(err)
-	}
+	k.cdc.MustUnmarshalBinary(b, &candidate)
 	return candidate, true
 }
 
 // Get the set of all candidates, retrieve a maxRetrieve number of records
 func (k Keeper) GetCandidates(ctx sdk.Context, maxRetrieve int16) (candidates Candidates) {
 	store := ctx.KVStore(k.storeKey)
-	iterator := store.Iterator(subspace(CandidatesKey))
+	iterator := store.SubspaceIterator(CandidatesKey)
 
 	candidates = make([]Candidate, maxRetrieve)
 	i := 0
@@ -88,10 +81,7 @@ func (k Keeper) GetCandidates(ctx sdk.Context, maxRetrieve int16) (candidates Ca
 		}
 		bz := iterator.Value()
 		var candidate Candidate
-		err := k.cdc.UnmarshalBinary(bz, &candidate)
-		if err != nil {
-			panic(err)
-		}
+		k.cdc.MustUnmarshalBinary(bz, &candidate)
 		candidates[i] = candidate
 		iterator.Next()
 	}
@@ -112,73 +102,48 @@ func (k Keeper) setCandidate(ctx sdk.Context, candidate Candidate) {
 	}
 
 	// marshal the candidate record and add to the state
-	bz, err := k.cdc.MarshalBinary(candidate)
-	if err != nil {
-		panic(err)
-	}
-	store.Set(GetCandidateKey(candidate.Address), bz)
+	bz := k.cdc.MustMarshalBinary(candidate)
+	store.Set(GetCandidateKey(address), bz)
 
-	// if the voting power is the same no need to update any of the other indexes
-	if oldFound && oldCandidate.Assets.Equal(candidate.Assets) {
+	if oldFound {
+		// if the voting power is the same no need to update any of the other indexes
+		if oldCandidate.BondedShares.Equal(candidate.BondedShares) {
+			return
+		}
+
+		// if this candidate wasn't just bonded then update the height and counter
+		if oldCandidate.Status != Bonded {
+			candidate.ValidatorBondHeight = ctx.BlockHeight()
+			counter := k.getIntraTxCounter(ctx)
+			candidate.ValidatorBondCounter = counter
+			k.setIntraTxCounter(ctx, counter+1)
+		}
+
+		// delete the old record in the power ordered list
+		store.Delete(GetValidatorKey(oldCandidate.validator()))
+	}
+
+	// set the new candidate record
+	bz = k.cdc.MustMarshalBinary(candidate)
+	store.Set(GetCandidateKey(address), bz)
+
+	// update the list ordered by voting power
+	validator := candidate.validator()
+	bzVal := k.cdc.MustMarshalBinary(validator)
+	store.Set(GetValidatorKey(validator), bzVal)
+
+	// add to the validators to update list if is already a validator
+	if store.Get(GetRecentValidatorKey(candidate.PubKey)) != nil {
+		bzAbci := k.cdc.MustMarshalBinary(validator.abciValidator(k.cdc))
+		store.Set(GetAccUpdateValidatorKey(address), bzAbci)
+
+		// also update the recent validator store
+		store.Set(GetRecentValidatorKey(validator.PubKey), bzVal)
 		return
 	}
 
-	updateHeight := false
-
-	// update the list ordered by voting power
-	if oldFound {
-		if !k.isNewValidator(ctx, store, candidate.Address) {
-			updateHeight = true
-		}
-		// else already in the validator set - retain the old validator height and counter
-		store.Delete(GetValidatorKey(address, oldCandidate.Assets, oldCandidate.ValidatorBondHeight, oldCandidate.ValidatorBondCounter, k.cdc))
-	} else {
-		updateHeight = true
-	}
-
-	if updateHeight {
-		// wasn't a candidate or wasn't in the validator set, update the validator block height and counter
-		candidate.ValidatorBondHeight = ctx.BlockHeight()
-		counter := k.getCounter(ctx)
-		candidate.ValidatorBondCounter = counter
-		k.setCounter(ctx, counter+1)
-	}
-
-	// update the candidate record
-	bz, err = k.cdc.MarshalBinary(candidate)
-	if err != nil {
-		panic(err)
-	}
-	store.Set(GetCandidateKey(candidate.Address), bz)
-
-	// marshal the new validator record
-	validator := candidate.validator()
-	bz, err = k.cdc.MarshalBinary(validator)
-	if err != nil {
-		panic(err)
-	}
-
-	store.Set(GetValidatorKey(address, validator.Power, validator.Height, validator.Counter, k.cdc), bz)
-
-	// add to the validators to update list if is already a validator
-	// or is a new validator
-	setAcc := false
-	if store.Get(GetRecentValidatorKey(address)) != nil {
-		setAcc = true
-
-		// want to check in the else statement because inefficient
-	} else if k.isNewValidator(ctx, store, address) {
-		setAcc = true
-	}
-	if setAcc {
-		bz, err = k.cdc.MarshalBinary(validator.abciValidator(k.cdc))
-		if err != nil {
-			panic(err)
-		}
-		store.Set(GetAccUpdateValidatorKey(validator.Address), bz)
-
-	}
-
+	// maybe add to the validator list and kick somebody off
+	k.addNewValidatorOrNot(ctx, store, candidate.Address)
 	return
 }
 
@@ -193,36 +158,78 @@ func (k Keeper) removeCandidate(ctx sdk.Context, address sdk.Address) {
 	// delete the old candidate record
 	store := ctx.KVStore(k.storeKey)
 	store.Delete(GetCandidateKey(address))
-	store.Delete(GetValidatorKey(address, candidate.Assets, candidate.ValidatorBondHeight, candidate.ValidatorBondCounter, k.cdc))
+	store.Delete(GetValidatorKey(candidate.validator()))
 
 	// delete from recent and power weighted validator groups if the validator
 	// exists and add validator with zero power to the validator updates
-	if store.Get(GetRecentValidatorKey(address)) == nil {
+	if store.Get(GetRecentValidatorKey(candidate.PubKey)) == nil {
 		return
 	}
-	bz, err := k.cdc.MarshalBinary(candidate.validator().abciValidatorZero(k.cdc))
-	if err != nil {
-		panic(err)
-	}
+	bz := k.cdc.MustMarshalBinary(candidate.validator().abciValidatorZero(k.cdc))
 	store.Set(GetAccUpdateValidatorKey(address), bz)
-	store.Delete(GetRecentValidatorKey(address))
+	store.Delete(GetRecentValidatorKey(candidate.PubKey))
 }
 
 //___________________________________________________________________________
 
-// Get the validator set from the candidates. The correct subset is retrieved
-// by iterating through an index of the candidates sorted by power, stored
-// using the ValidatorsKey. Simultaniously the most recent the validator
-// records are updated in store with the RecentValidatorsKey. This store is
-// used to determine if a candidate is a validator without needing to iterate
-// over the subspace as we do in GetValidators
+// get the group of the most recent validators
 func (k Keeper) GetValidators(ctx sdk.Context) (validators []Validator) {
 	store := ctx.KVStore(k.storeKey)
 
-	// clear the recent validators store, add to the ToKickOut Temp store
-	iterator := store.Iterator(subspace(RecentValidatorsKey))
+	// add the actual validator power sorted store
+	maxValidators := k.GetParams(ctx).MaxValidators
+	validators = make([]Validator, maxValidators)
+
+	iterator := store.SubspaceIterator(RecentValidatorsKey)
+	i := 0
 	for ; iterator.Valid(); iterator.Next() {
-		addr := AddrFromKey(iterator.Key())
+		bz := iterator.Value()
+		var validator Validator
+		k.cdc.MustUnmarshalBinary(bz, &validator)
+		validators[i] = validator
+		i++
+	}
+	iterator.Close()
+	return validators[:i] // trim
+}
+
+// Only used for testing
+// get the group of the most recent validators
+func (k Keeper) getValidatorsOrdered(ctx sdk.Context) []Validator {
+	vals := k.GetValidators(ctx)
+	sort.Sort(sort.Reverse(validators(vals)))
+	return vals
+}
+
+// Is the address provided a part of the most recently saved validator group?
+func (k Keeper) IsValidator(ctx sdk.Context, pk crypto.PubKey) bool {
+	store := ctx.KVStore(k.storeKey)
+	if store.Get(GetRecentValidatorKey(pk)) == nil {
+		return false
+	}
+	return true
+}
+
+// This function add's (or doesn't add) a candidate record to the validator group
+// simultaniously it kicks any old validators out
+//
+// The correct subset is retrieved by iterating through an index of the
+// candidates sorted by power, stored using the ValidatorsKey. Simultaniously
+// the most recent the validator records are updated in store with the
+// RecentValidatorsKey. This store is used to determine if a candidate is a
+// validator without needing to iterate over the subspace as we do in
+// GetValidators
+func (k Keeper) addNewValidatorOrNot(ctx sdk.Context, store sdk.KVStore, address sdk.Address) {
+
+	// clear the recent validators store, add to the ToKickOut temp store
+	iterator := store.SubspaceIterator(RecentValidatorsKey)
+	for ; iterator.Valid(); iterator.Next() {
+
+		bz := iterator.Value()
+		var validator Validator
+		k.cdc.MustUnmarshalBinary(bz, &validator)
+
+		addr := validator.Address
 
 		// iterator.Value is the validator object
 		store.Set(GetToKickOutValidatorKey(addr), iterator.Value())
@@ -232,8 +239,7 @@ func (k Keeper) GetValidators(ctx sdk.Context) (validators []Validator) {
 
 	// add the actual validator power sorted store
 	maxValidators := k.GetParams(ctx).MaxValidators
-	iterator = store.ReverseIterator(subspace(ValidatorsKey)) // largest to smallest
-	validators = make([]Validator, maxValidators)
+	iterator = store.ReverseSubspaceIterator(ValidatorsKey) // largest to smallest
 	i := 0
 	for ; ; i++ {
 		if !iterator.Valid() || i > int(maxValidators-1) {
@@ -242,23 +248,25 @@ func (k Keeper) GetValidators(ctx sdk.Context) (validators []Validator) {
 		}
 		bz := iterator.Value()
 		var validator Validator
-		err := k.cdc.UnmarshalBinary(bz, &validator)
-		if err != nil {
-			panic(err)
-		}
-		validators[i] = validator
+		k.cdc.MustUnmarshalBinary(bz, &validator)
 
 		// remove from ToKickOut group
 		store.Delete(GetToKickOutValidatorKey(validator.Address))
 
 		// also add to the recent validators group
-		store.Set(GetRecentValidatorKey(validator.Address), bz)
+		store.Set(GetRecentValidatorKey(validator.PubKey), bz)
+
+		// MOST IMPORTANTLY, add to the accumulated changes if this is the modified candidate
+		if bytes.Equal(address, validator.Address) {
+			bz = k.cdc.MustMarshalBinary(validator.abciValidator(k.cdc))
+			store.Set(GetAccUpdateValidatorKey(address), bz)
+		}
 
 		iterator.Next()
 	}
 
 	// add any kicked out validators to the acc change
-	iterator = store.Iterator(subspace(ToKickOutValidatorsKey))
+	iterator = store.SubspaceIterator(ToKickOutValidatorsKey)
 	for ; iterator.Valid(); iterator.Next() {
 		key := iterator.Key()
 		addr := AddrFromKey(key)
@@ -266,57 +274,52 @@ func (k Keeper) GetValidators(ctx sdk.Context) (validators []Validator) {
 		// get the zero abci validator from the ToKickOut iterator value
 		bz := iterator.Value()
 		var validator Validator
-		err := k.cdc.UnmarshalBinary(bz, &validator)
-		if err != nil {
-			panic(err)
-		}
-		bz, err = k.cdc.MarshalBinary(validator.abciValidatorZero(k.cdc))
-		if err != nil {
-			panic(err)
-		}
+		k.cdc.MustUnmarshalBinary(bz, &validator)
+		bz = k.cdc.MustMarshalBinary(validator.abciValidatorZero(k.cdc))
 
 		store.Set(GetAccUpdateValidatorKey(addr), bz)
 		store.Delete(key)
 	}
 	iterator.Close()
-
-	return validators[:i] // trim
 }
 
-// TODO this is madly inefficient because need to call every time we set a candidate
-// Should use something better than an iterator maybe?
-// Used to determine if something has just been added to the actual validator set
-func (k Keeper) isNewValidator(ctx sdk.Context, store sdk.KVStore, address sdk.Address) bool {
-	// add the actual validator power sorted store
-	maxVal := k.GetParams(ctx).MaxValidators
-	iterator := store.ReverseIterator(subspace(ValidatorsKey)) // largest to smallest
-	for i := 0; ; i++ {
-		if !iterator.Valid() || i > int(maxVal-1) {
-			iterator.Close()
-			break
-		}
-		bz := iterator.Value()
-		var val Validator
-		err := k.cdc.UnmarshalBinary(bz, &val)
-		if err != nil {
-			panic(err)
-		}
-		if bytes.Equal(val.Address, address) {
-			return true
-		}
-		iterator.Next()
-	}
-
-	return false
-}
-
-// Is the address provided a part of the most recently saved validator group?
-func (k Keeper) IsRecentValidator(ctx sdk.Context, address sdk.Address) bool {
+// cummulative power of the non-absent prevotes
+func (k Keeper) GetTotalPrecommitVotingPower(ctx sdk.Context) sdk.Rat {
 	store := ctx.KVStore(k.storeKey)
-	if store.Get(GetRecentValidatorKey(address)) == nil {
-		return false
+
+	// get absent prevote indexes
+	absents := ctx.AbsentValidators()
+
+	TotalPower := sdk.ZeroRat()
+	i := int32(0)
+	iterator := store.SubspaceIterator(RecentValidatorsKey)
+	for ; iterator.Valid(); iterator.Next() {
+
+		skip := false
+		for j, absentIndex := range absents {
+			if absentIndex > i {
+				break
+			}
+
+			// if non-voting validator found, skip adding its power
+			if absentIndex == i {
+				absents = append(absents[:j], absents[j+1:]...) // won't need again
+				skip = true
+				break
+			}
+		}
+		if skip {
+			continue
+		}
+
+		bz := iterator.Value()
+		var validator Validator
+		k.cdc.MustUnmarshalBinary(bz, &validator)
+		TotalPower = TotalPower.Add(validator.Power)
+		i++
 	}
-	return true
+	iterator.Close()
+	return TotalPower
 }
 
 //_________________________________________________________________________
@@ -326,14 +329,11 @@ func (k Keeper) IsRecentValidator(ctx sdk.Context, address sdk.Address) bool {
 func (k Keeper) getAccUpdateValidators(ctx sdk.Context) (updates []abci.Validator) {
 	store := ctx.KVStore(k.storeKey)
 
-	iterator := store.Iterator(subspace(AccUpdateValidatorsKey)) //smallest to largest
+	iterator := store.SubspaceIterator(AccUpdateValidatorsKey) //smallest to largest
 	for ; iterator.Valid(); iterator.Next() {
 		valBytes := iterator.Value()
 		var val abci.Validator
-		err := k.cdc.UnmarshalBinary(valBytes, &val)
-		if err != nil {
-			panic(err)
-		}
+		k.cdc.MustUnmarshalBinary(valBytes, &val)
 		updates = append(updates, val)
 	}
 	iterator.Close()
@@ -345,7 +345,7 @@ func (k Keeper) clearAccUpdateValidators(ctx sdk.Context) {
 	store := ctx.KVStore(k.storeKey)
 
 	// delete subspace
-	iterator := store.Iterator(subspace(AccUpdateValidatorsKey))
+	iterator := store.SubspaceIterator(AccUpdateValidatorsKey)
 	for ; iterator.Valid(); iterator.Next() {
 		store.Delete(iterator.Key())
 	}
@@ -364,17 +364,14 @@ func (k Keeper) GetDelegatorBond(ctx sdk.Context,
 		return bond, false
 	}
 
-	err := k.cdc.UnmarshalBinary(delegatorBytes, &bond)
-	if err != nil {
-		panic(err)
-	}
+	k.cdc.MustUnmarshalBinary(delegatorBytes, &bond)
 	return bond, true
 }
 
 // load all bonds
 func (k Keeper) getBonds(ctx sdk.Context, maxRetrieve int16) (bonds []DelegatorBond) {
 	store := ctx.KVStore(k.storeKey)
-	iterator := store.Iterator(subspace(DelegatorBondKeyPrefix))
+	iterator := store.SubspaceIterator(DelegatorBondKeyPrefix)
 
 	bonds = make([]DelegatorBond, maxRetrieve)
 	i := 0
@@ -385,10 +382,7 @@ func (k Keeper) getBonds(ctx sdk.Context, maxRetrieve int16) (bonds []DelegatorB
 		}
 		bondBytes := iterator.Value()
 		var bond DelegatorBond
-		err := k.cdc.UnmarshalBinary(bondBytes, &bond)
-		if err != nil {
-			panic(err)
-		}
+		k.cdc.MustUnmarshalBinary(bondBytes, &bond)
 		bonds[i] = bond
 		iterator.Next()
 	}
@@ -399,7 +393,7 @@ func (k Keeper) getBonds(ctx sdk.Context, maxRetrieve int16) (bonds []DelegatorB
 func (k Keeper) GetDelegatorBonds(ctx sdk.Context, delegator sdk.Address, maxRetrieve int16) (bonds []DelegatorBond) {
 	store := ctx.KVStore(k.storeKey)
 	delegatorPrefixKey := GetDelegatorBondsKey(delegator, k.cdc)
-	iterator := store.Iterator(subspace(delegatorPrefixKey)) //smallest to largest
+	iterator := store.SubspaceIterator(delegatorPrefixKey) //smallest to largest
 
 	bonds = make([]DelegatorBond, maxRetrieve)
 	i := 0
@@ -410,10 +404,7 @@ func (k Keeper) GetDelegatorBonds(ctx sdk.Context, delegator sdk.Address, maxRet
 		}
 		bondBytes := iterator.Value()
 		var bond DelegatorBond
-		err := k.cdc.UnmarshalBinary(bondBytes, &bond)
-		if err != nil {
-			panic(err)
-		}
+		k.cdc.MustUnmarshalBinary(bondBytes, &bond)
 		bonds[i] = bond
 		iterator.Next()
 	}
@@ -422,16 +413,42 @@ func (k Keeper) GetDelegatorBonds(ctx sdk.Context, delegator sdk.Address, maxRet
 
 func (k Keeper) setDelegatorBond(ctx sdk.Context, bond DelegatorBond) {
 	store := ctx.KVStore(k.storeKey)
-	b, err := k.cdc.MarshalBinary(bond)
-	if err != nil {
-		panic(err)
-	}
+	b := k.cdc.MustMarshalBinary(bond)
 	store.Set(GetDelegatorBondKey(bond.DelegatorAddr, bond.CandidateAddr, k.cdc), b)
 }
 
 func (k Keeper) removeDelegatorBond(ctx sdk.Context, bond DelegatorBond) {
 	store := ctx.KVStore(k.storeKey)
 	store.Delete(GetDelegatorBondKey(bond.DelegatorAddr, bond.CandidateAddr, k.cdc))
+}
+
+//_______________________________________________________________________
+
+// XXX TODO trim functionality
+
+// retrieve all the power changes which occur after a height
+func (k Keeper) GetPowerChangesAfterHeight(ctx sdk.Context, earliestHeight int64) (pcs []PowerChange) {
+	store := ctx.KVStore(k.storeKey)
+
+	iterator := store.SubspaceIterator(PowerChangeKey) //smallest to largest
+	for ; iterator.Valid(); iterator.Next() {
+		pcBytes := iterator.Value()
+		var pc PowerChange
+		k.cdc.MustUnmarshalBinary(pcBytes, &pc)
+		if pc.Height < earliestHeight {
+			break
+		}
+		pcs = append(pcs, pc)
+	}
+	iterator.Close()
+	return
+}
+
+// set a power change
+func (k Keeper) setPowerChange(ctx sdk.Context, pc PowerChange) {
+	store := ctx.KVStore(k.storeKey)
+	b := k.cdc.MustMarshalBinary(pc)
+	store.Set(GetPowerChangeKey(pc.Height), b)
 }
 
 //_______________________________________________________________________
@@ -448,18 +465,12 @@ func (k Keeper) GetParams(ctx sdk.Context) (params Params) {
 		panic("Stored params should not have been nil")
 	}
 
-	err := k.cdc.UnmarshalBinary(b, &params)
-	if err != nil {
-		panic(err)
-	}
+	k.cdc.MustUnmarshalBinary(b, &params)
 	return
 }
 func (k Keeper) setParams(ctx sdk.Context, params Params) {
 	store := ctx.KVStore(k.storeKey)
-	b, err := k.cdc.MarshalBinary(params)
-	if err != nil {
-		panic(err)
-	}
+	b := k.cdc.MustMarshalBinary(params)
 	store.Set(ParamKey, b)
 	k.params = Params{} // clear the cache
 }
@@ -477,19 +488,13 @@ func (k Keeper) GetPool(ctx sdk.Context) (pool Pool) {
 	if b == nil {
 		panic("Stored pool should not have been nil")
 	}
-	err := k.cdc.UnmarshalBinary(b, &pool)
-	if err != nil {
-		panic(err) // This error should never occur big problem if does
-	}
+	k.cdc.MustUnmarshalBinary(b, &pool)
 	return
 }
 
 func (k Keeper) setPool(ctx sdk.Context, p Pool) {
 	store := ctx.KVStore(k.storeKey)
-	b, err := k.cdc.MarshalBinary(p)
-	if err != nil {
-		panic(err)
-	}
+	b := k.cdc.MustMarshalBinary(p)
 	store.Set(PoolKey, b)
 	k.pool = Pool{} //clear the cache
 }